removed inconsistency in FormWindow.SetContent.
There is still a competition between the WMComponents.FormWindow.SetContent process and the WindowManager.AddWindow process, leading to occasional (timing dependent?) non-display of the contents of a formwindow. this inconsistancy is probably due to the sequencer exchange when SetContent is called.
However, this seems mainly to matter when empty formwindows are put to the display space and only then, SetContent is called. Should be eliminated anyway.